requirements.txt
fastapi
uvicorn
mcp
requests
google-generativeai
